unit testing(jest)

+++ src/movies/movies.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { Movie } from './entities/movie.entity';
 import { CreateMovieDto } from './dto/create-movie.dto';
 import { UpdateMovieDto } from './dto/update-movie.dto';
@@ -12,11 +12,19 @@ export class MoviesService {
   }
 
   getOne(id: number): Movie {
-    return this.movies.find((movie) => movie.id === id);
+    const movie = this.movies.find((movie) => movie.id === id);
+    if (!movie) {
+      throw new NotFoundException(`movie with id ${id} not found`);
+    }
+    return movie;
   }
 
   deleteOne(id: number): boolean {
-    this.movies.filter((movie) => movie.id !== id);
+    const movie = this.getOne(id);
+    if (!movie) {
+      throw new NotFoundException(`movie with id ${id} not found`);
+    }
+    this.movies = this.movies.filter((movie) => movie.id !== id);
     return true;
   }
 
@@ -30,6 +38,9 @@ export class MoviesService {
   //이름에 상관없이, arg의 순서에 따라 결정
   update(id: number, updatedData: UpdateMovieDto) {
     const movie = this.getOne(id);
+    if (!movie) {
+      throw new NotFoundException(`movie with id ${id} not found`);
+    }
     this.deleteOne(id);
     this.movies.push({ ...movie, ...updatedData });
   }
